feat: Add Frame enum
